+++ framework/CHANGELOG.md
@@ -80,6 +80,7 @@ Yii Framework 2 Change Log
 - Enh #12440: Added `yii\base\Event::offAll()` method allowing clear all registered class-level event handlers (klimov-paul)
 - Enh #12580: Make `yii.js` comply with strict and non-strict javascript mode to allow concatenation with external code (mikehaertl)
 - Enh: Method `yii\console\controllers\AssetController::getAssetManager()` automatically enables `yii\web\AssetManager::forceCopy` in case it is not explicitly specified (pana1990, klimov-paul)
+- Enh #11309: Added Method `yii\web\Request::getDomain()` which returns the domain without schema and port (ivanovyordan)
 
 
 2.0.9 July 11, 2016